rastercalc rxr werkend met storage
